Re-add the getAdmin() function at User model
This was deleted when merging.
Also change integer's to bigInteger's in the support request migration.

+++ database/migrations/2019_12_02_112121_create_support_requests_table.php
@@ -15,8 +15,8 @@ class CreateSupportRequestsTable extends Migration
     {
         Schema::create('support_requests', function (Blueprint $table) {
             $table->increments('id')->autoIncrement();
-            $table->integer('challenge_id')->unsigned();
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('challenge_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->text('subject');
             $table->longText('message');
             $table->boolean('solved')->default(false);
